+++ commands/image-editing.js
@@ -6,10 +6,6 @@ const Command = require("../structures/command.js"),
 
 Canvas.registerFont("assets/Oswald-Regular.ttf", {family: "Oswald"});
 
-function getPixelFactor(img) {
-	return Math.ceil(img.bitmap.width > img.bitmap.height ? img.bitmap.width : img.bitmap.height) / 100;
-}
-
 module.exports = [
 	class BlurCommand extends Command {
 		constructor() {
@@ -42,31 +38,16 @@ module.exports = [
 					user: [],
 					level: 0
 				},
-				usage: "blur [image URL or mention] [--level <number>]"
+				usage: "blur [image URL/mention/emoji] [--level <number>]"
 			});
 		}
 		
 		async run(bot, message, args, flags) {
-			let imageURL = args[0];
-			if (!imageURL) {
-				imageURL = await imageManager.resolveImageURL(message);
-				if (!imageURL) return {cmdWarn: "No image attachment found in recent messages"};
-			}
+			const imgResolvable = await imageManager.getImageResolvable(message, args[0]);
+			if (!imgResolvable) return {cmdWarn: "No mention or emoji found, or image attachment found in recent messages"};
 			
-			Jimp.read(imageURL)
-				.then(img => {
-					const levelFlag = flags.find(f => f.name == "level");
-					let blurLevel;
-					if (levelFlag) {
-						blurLevel = levelFlag.args;
-					} else {
-						blurLevel = getPixelFactor(img);
-					}
-					imageManager.postImage(message, img.blur(blurLevel), "blur.png");
-				})
-				.catch(() => {
-					message.channel.send("⚠ Failed to get image for that URL.");
-				});
+			const levelFlag = flags.find(f => f.name == "level");
+			imageManager.applyJimpFilterAndPost(message, imgResolvable, "blur", {blur: levelFlag ? levelFlag.args : null});
 		}
 	},
 	class CreateMemeCommand extends Command {
@@ -107,8 +88,8 @@ module.exports = [
 		}
 		
 		async run(bot, message, args, flags) {			
-			const imageURL = args[0] || await imageManager.resolveImageURL(message);
-			if (!imageURL) return {cmdWarn: "No image attachment found in recent messages"};
+			const imgResolvable = await imageManager.getImageResolvable(message, args[0]);
+			if (!imgResolvable) return {cmdWarn: "No mention or emoji found, or image attachment found in recent messages"};
 
 			const pipeRegex = / ?\| /,
 				disableCapsFlag = flags.some(f => f.name == "disablecaps");
@@ -166,7 +147,7 @@ module.exports = [
 				});
 			};
 
-			imageManager.getCanvasImage(img, imageURL);
+			imageManager.getCanvasImage(img, imgResolvable, args[0] && args[0].isEmoji);
 		}
 
 		drawText(canvas, ctx, text, isTop) {
@@ -233,24 +214,15 @@ module.exports = [
 					user: [],
 					level: 0
 				},
-				usage: "flip [image URL or mention]"
+				usage: "flip [image URL/mention/emoji]"
 			});
 		}
 		
 		async run(bot, message, args, flags) {
-			let imageURL = args[0];
-			if (!imageURL) {
-				imageURL = await imageManager.resolveImageURL(message);
-				if (!imageURL) return {cmdWarn: "No image attachment found in recent messages"};
-			}
+			const imgResolvable = await imageManager.getImageResolvable(message, args[0]);
+			if (!imgResolvable) return {cmdWarn: "No mention or emoji found, or image attachment found in recent messages"};
 			
-			Jimp.read(imageURL)
-				.then(img => {
-					imageManager.postImage(message, img.mirror(true, false), "flip.png");
-				})
-				.catch(() => {
-					message.channel.send("⚠ Failed to get image for that URL.");
-				});
+			imageManager.applyJimpFilterAndPost(message, imgResolvable, "flip");
 		}
 	},
 	class FlopCommand extends Command {
@@ -274,24 +246,15 @@ module.exports = [
 					user: [],
 					level: 0
 				},
-				usage: "flop [image URL or mention]"
+				usage: "flop [image URL/mention/emoji]"
 			});
 		}
 		
 		async run(bot, message, args, flags) {
-			let imageURL = args[0];
-			if (!imageURL) {
-				imageURL = await imageManager.resolveImageURL(message);
-				if (!imageURL) return {cmdWarn: "No image attachment found in recent messages"};
-			}
+			const imgResolvable = await imageManager.getImageResolvable(message, args[0]);
+			if (!imgResolvable) return {cmdWarn: "No mention or emoji found, or image attachment found in recent messages"};
 			
-			Jimp.read(imageURL)
-				.then(img => {
-					imageManager.postImage(message, img.mirror(false, true), "flop.png");
-				})
-				.catch(() => {
-					message.channel.send("⚠ Failed to get image for that URL.");
-				});
+			imageManager.applyJimpFilterAndPost(message, imgResolvable, "flop");
 		}
 	},
 	class GrayscaleCommand extends Command {
@@ -316,24 +279,15 @@ module.exports = [
 					user: [],
 					level: 0
 				},
-				usage: "grayscale [image URL or mention]"
+				usage: "grayscale [image URL/mention/emoji]"
 			});
 		}
 		
 		async run(bot, message, args, flags) {
-			let imageURL = args[0];
-			if (!imageURL) {
-				imageURL = await imageManager.resolveImageURL(message);
-				if (!imageURL) return {cmdWarn: "No image attachment found in recent messages"};
-			}
+			const imgResolvable = await imageManager.getImageResolvable(message, args[0]);
+			if (!imgResolvable) return {cmdWarn: "No mention or emoji found, or image attachment found in recent messages"};
 			
-			Jimp.read(imageURL)
-				.then(img => {
-					imageManager.postImage(message, img.grayscale(), "grayscale.png");
-				})
-				.catch(() => {
-					message.channel.send("⚠ Failed to get image for that URL.");
-				});
+			imageManager.applyJimpFilterAndPost(message, imgResolvable, "grayscale");
 		}
 	},
 	class InvertCommand extends Command {
@@ -357,24 +311,15 @@ module.exports = [
 					user: [],
 					level: 0
 				},
-				usage: "invert [image URL or mention]"
+				usage: "invert [image URL/mention/emoji]"
 			});
 		}
 		
 		async run(bot, message, args, flags) {
-			let imageURL = args[0];
-			if (!imageURL) {
-				imageURL = await imageManager.resolveImageURL(message);
-				if (!imageURL) return {cmdWarn: "No image attachment found in recent messages"};
-			}
+			const imgResolvable = await imageManager.getImageResolvable(message, args[0]);
+			if (!imgResolvable) return {cmdWarn: "No mention or emoji found, or image attachment found in recent messages"};
 			
-			Jimp.read(imageURL)
-				.then(img => {
-					imageManager.postImage(message, img.invert(), "invert.png");
-				})
-				.catch(() => {
-					message.channel.send("⚠ Failed to get image for that URL.");
-				});
+			imageManager.applyJimpFilterAndPost(message, imgResolvable, "invert");
 		}
 	},
 	class MirrorCommand extends Command {
@@ -402,20 +347,16 @@ module.exports = [
 					user: [],
 					level: 0
 				},
-				usage: "mirror [image URL or mention] <[haah | right-to-left] | [hooh | bottom-to-top] | [waaw | left-to-right] | [woow | top-to-bottom]>"
+				usage: "mirror [image URL/mention/emoji] <[haah | right-to-left] | [hooh | bottom-to-top] | [waaw | left-to-right] | [woow | top-to-bottom]>"
 			});
 		}
 		
 		async run(bot, message, args, flags) {
+			const imgResolvable = await imageManager.getImageResolvable(message, args[0]);
+			if (!imgResolvable) return {cmdWarn: "No mention or emoji found, or image attachment found in recent messages"};
+			
 			const type = args[1];
-			let imageURL = args[0];
-			
-			if (!imageURL) {
-				imageURL = await imageManager.resolveImageURL(message);
-				if (!imageURL) return {cmdWarn: "No image attachment found in recent messages"};
-			}
-			
-			Jimp.read(imageURL)
+			Jimp.read(imgResolvable)
 				.then(img => {
 					const imgClone1 = img.clone(),
 						imgClone2 = img.clone(),
@@ -430,7 +371,7 @@ module.exports = [
 						new Jimp(imgWidth, imgHeight, (err, img2) => {
 							img2.composite(imgClone1, imgWidth / 2, 0)
 								.composite(imgClone2, 0, 0);
-							imageManager.postImage(message, img2, "mirror-haah.png");
+							imageManager.postJimpImage(message, img2, "mirror-haah.png");
 						});
 						return;
 					} else if (type == "hooh" || type == "bottom-to-top") {
@@ -441,7 +382,7 @@ module.exports = [
 						new Jimp(imgWidth, imgHeight, (err, img2) => {
 							img2.composite(imgClone1, 0, imgHeight / 2)
 								.composite(imgClone2, 0, 0);
-							imageManager.postImage(message, img2, "mirror-hooh.png");
+							imageManager.postJimpImage(message, img2, "mirror-hooh.png");
 						});
 					} else if (type == "waaw" || type == "left-to-right") {
 						imgClone1.crop(0, 0, imgWidth / 2, imgHeight);
@@ -451,7 +392,7 @@ module.exports = [
 						new Jimp(imgWidth, imgHeight, (err, img2) => {
 							img2.composite(imgClone1, 0, 0)
 								.composite(imgClone2, imgWidth / 2, 0);
-							imageManager.postImage(message, img2, "mirror-waaw.png");
+							imageManager.postJimpImage(message, img2, "mirror-waaw.png");
 						});
 					} else {
 						imgClone1.crop(0, 0, imgWidth, imgHeight / 2);
@@ -461,7 +402,7 @@ module.exports = [
 						new Jimp(imgWidth, imgHeight, (err, img2) => {
 							img2.composite(imgClone1, 0, 0)
 								.composite(imgClone2, 0, imgHeight / 2);
-							imageManager.postImage(message, img2, "mirror-woow.png");
+							imageManager.postJimpImage(message, img2, "mirror-woow.png");
 						});
 					}
 				})
@@ -492,27 +433,21 @@ module.exports = [
 					user: [],
 					level: 0
 				},
-				usage: "needsmorejpeg [image URL or mention]"
+				usage: "needsmorejpeg [image URL/mention/emoji]"
 			});
 		}
 		
 		async run(bot, message, args, flags) {
-			let imageURL = args[0];
-			if (!imageURL) {
-				imageURL = await imageManager.resolveImageURL(message);
-				if (!imageURL) return {cmdWarn: "No image attachment found in recent messages"};
-			}
+			const imgResolvable = await imageManager.getImageResolvable(message, args[0]);
+			if (!imgResolvable) return {cmdWarn: "No mention or emoji found, or image attachment found in recent messages"};
 			
-			Jimp.read(imageURL)
+			Jimp.read(imgResolvable)
 				.then(img => {
 					img.quality(1)
 						.getBufferAsync(Jimp.MIME_JPEG)
 						.then(imgToSend => {
 							message.channel.send({
-								files: [{
-									attachment: imgToSend,
-									name: "needsmorejpeg.jpg"
-								}]
+								files: [{attachment: imgToSend, name: "needsmorejpeg.jpg"}]
 							});
 						})
 						.catch(() => {
@@ -541,29 +476,31 @@ module.exports = [
 					time: 15000,
 					type: "channel"
 				},
+				flags: [
+					{
+						name: "pixels",
+						desc: "The width of each enlarged pixel",
+						arg: {
+							type: "number",
+							min: 1
+						}
+					}
+				],
 				perms: {
 					bot: ["ATTACH_FILES"],
 					user: [],
 					level: 0
 				},
-				usage: "pixelate [image URL or mention]"
+				usage: "pixelate [image URL/mention/emoji] [--pixels <number>]"
 			});
 		}
 		
 		async run(bot, message, args, flags) {
-			let imageURL = args[0];
-			if (!imageURL) {
-				imageURL = await imageManager.resolveImageURL(message);
-				if (!imageURL) return {cmdWarn: "No image attachment found in recent messages"};
-			}
+			const imgResolvable = await imageManager.getImageResolvable(message, args[0]);
+			if (!imgResolvable) return {cmdWarn: "No mention or emoji found, or image attachment found in recent messages"};
 			
-			Jimp.read(imageURL)
-				.then(img => {
-					imageManager.postImage(message, img.pixelate(getPixelFactor(img)), "pixelate.png");
-				})
-				.catch(() => {
-					message.channel.send("⚠ Failed to get image for that URL.");
-				});
+			const pixelsFlag = flags.find(f => f.name == "pixels");
+			imageManager.applyJimpFilterAndPost(message, imgResolvable, "pixelate", {pixels: pixelsFlag ? pixelsFlag.args : null});
 		}
 	},
 	class RotateCommand extends Command {
@@ -598,25 +535,16 @@ module.exports = [
 					user: [],
 					level: 0
 				},
-				usage: "rotate [image URL or mention] [--degrees <1-359>]"
+				usage: "rotate [image URL/mention/emoji] [--degrees <1-359>]"
 			});
 		}
 		
 		async run(bot, message, args, flags) {
-			let imageURL = args[0];
-			if (!imageURL) {
-				imageURL = await imageManager.resolveImageURL(message);
-				if (!imageURL) return {cmdWarn: "No image attachment found in recent messages"};
-			}
+			const imgResolvable = await imageManager.getImageResolvable(message, args[0]);
+			if (!imgResolvable) return {cmdWarn: "No mention or emoji found, or image attachment found in recent messages"};
 			
-			Jimp.read(imageURL)
-				.then(img => {
-					const levelFlag = flags.find(f => f.name == "degrees");
-					imageManager.postImage(message, img.rotate(levelFlag ? levelFlag.args : 90), "rotate.png");
-				})
-				.catch(() => {
-					message.channel.send("⚠ Failed to get image for that URL.");
-				});
+			const degreesFlag = flags.find(f => f.name == "degrees");
+			imageManager.applyJimpFilterAndPost(message, imgResolvable, "rotate", {rotation: degreesFlag ? degreesFlag.args : null});
 		}
 	},
 	class SepiaCommand extends Command {
@@ -640,24 +568,15 @@ module.exports = [
 					user: [],
 					level: 0
 				},
-				usage: "sepia [image URL or mention]"
+				usage: "sepia [image URL/mention/emoji]"
 			});
 		}
 		
 		async run(bot, message, args, flags) {
-			let imageURL = args[0];
-			if (!imageURL) {
-				imageURL = await imageManager.resolveImageURL(message);
-				if (!imageURL) return {cmdWarn: "No image attachment found in recent messages"};
-			}
+			const imgResolvable = await imageManager.getImageResolvable(message, args[0]);
+			if (!imgResolvable) return {cmdWarn: "No mention or emoji found, or image attachment found in recent messages"};
 			
-			Jimp.read(imageURL)
-				.then(img => {
-					imageManager.postImage(message, img.sepia(), "sepia.png");
-				})
-				.catch(() => {
-					message.channel.send("⚠ Failed to get image for that URL.");
-				});
+			imageManager.applyJimpFilterAndPost(message, imgResolvable, "sepia");
 		}
 	},
 	class SpinCommand extends Command {
@@ -693,13 +612,13 @@ module.exports = [
 					user: [],
 					level: 0
 				},
-				usage: "spin [image URL or mention] [--speed <1-5>]"
+				usage: "spin [image URL/mention/emoji] [--speed <1-5>]"
 			});
 		}
 		
 		async run(bot, message, args, flags) {
-			const imageURL = args[0] || await imageManager.resolveImageUrl(message);
-			if (!imageURL) return {cmdWarn: "No image attachment found in recent messages"};
+			const imgResolvable = await imageManager.getImageResolvable(message, args[0]);
+			if (!imgResolvable) return {cmdWarn: "No mention or emoji found, or image attachment found in recent messages"};
 
 			const speedFlag = flags.find(f => f.name == "speed"),
 				img = new Canvas.Image();
@@ -753,7 +672,7 @@ module.exports = [
 					}]
 				});
 			};
-			imageManager.getCanvasImage(img, imageURL);
+			imageManager.getCanvasImage(img, imgResolvable, args[0] && args[0].isEmoji);
 		}
 	}
 ];