+++ utils/imageManager.js
@@ -1,47 +1,79 @@
-const Jimp = require("jimp"),
-	request = require("request");
+const {applyJimpFilter} = require("../modules/filters.js"),
+	Jimp = require("jimp"),
+	request = require("request"),
+	svg2png = require("svg2png");
 
 module.exports = {
-	resolveImageURL: async message => {
-		let imageURL = null;
-		await message.channel.fetchMessages({limit: 25})
-			.then(msgs => {
-				msgs = msgs.array();
-				for (const msg of msgs) {
-					if (msg.attachments.size > 0) {
-						imageURL = msg.attachments.last().url;
-						break;
-					} else {
-						const lastEmbed = msg.embeds[msg.embeds.length - 1];
-						if (lastEmbed && lastEmbed.image) {
-							imageURL = lastEmbed.image.url;
+	getImageResolvable: async (message, userInput) => {
+		let resolvedContent = null;
+		if (userInput && userInput.isEmoji) {
+			await svg2png(userInput.content, {width: 512, height: 512})
+				.then(pngBuffer => resolvedContent = Buffer.from(pngBuffer))
+				.catch(err => console.log(err));
+		} else if (!userInput) {
+			await message.channel.fetchMessages({limit: 25})
+				.then(msgs => {
+					msgs = msgs.array();
+					for (const msg of msgs) {
+						if (msg.attachments.size > 0) {
+							resolvedContent = msg.attachments.last().url;
 							break;
-						} else if (lastEmbed && lastEmbed.type == "image") {
-							imageURL = lastEmbed.url;
-							break;
+						} else {
+							const lastEmbed = msg.embeds[msg.embeds.length - 1];
+							if (lastEmbed && lastEmbed.image) {
+								resolvedContent = lastEmbed.image.url;
+								break;
+							} else if (lastEmbed && lastEmbed.type == "image") {
+								resolvedContent = lastEmbed.url;
+								break;
+							}
 						}
 					}
+				})
+				.catch(err => console.log("Failed to fetch messages while resolving an image URL:", err));
+		} else {
+			resolvedContent = userInput;
+		}
+		return resolvedContent;
+	},
+	getCanvasImage: (img, imgResolvable, isEmoji) => {
+		if (isEmoji) {
+			img.src = imgResolvable;
+		} else {
+			img.onerror = () => {
+				console.log("Failed to load the image.");
+			};
+			request.get({
+				url: imgResolvable,
+				encoding: null
+			}, (err, res) => {
+				if (err || !res || res.statusCode >= 400) {
+					console.log("Failed to get URL while trying to assign the source to the canvas image.");
+				} else {
+					img.src = res.body;
 				}
+			});
+		}
+	},
+	applyJimpFilterAndPost: (msg, imgResolvable, filter, options = {}) => {
+		Jimp.read(imgResolvable)
+			.then(img => {
+				applyJimpFilter(img, filter, options);
+				img.getBufferAsync(Jimp.MIME_PNG)
+					.then(imgToSend => {
+						msg.channel.send({
+							files: [{attachment: imgToSend, name: filter + ".png"}]
+						});
+					})
+					.catch(() => {
+						msg.channel.send("Failed to generate the image.");
+					});
 			})
-			.catch(err => console.log("Failed to fetch messages while resolving an image URL:", err));
-		return imageURL;
+			.catch(() => {
+				msg.channel.send("⚠ Failed to get image for that URL.");
+			});	
 	},
-	getCanvasImage: (img, url) => {
-		img.onerror = () => {
-			console.log("Failed to load the image.");
-		};
-		request.get({
-			url: url,
-			encoding: null
-		}, (err, res) => {
-			if (err || !res || res.statusCode >= 400) {
-				console.log("Failed to get URL while trying to assign the source to the canvas image.");
-			} else {
-				img.src = res.body;
-			}
-		});
-	},
-	postImage: (msg, img, fileName) => {
+	postJimpImage: (msg, img, fileName) => {
 		img.getBufferAsync(Jimp.MIME_PNG)
 			.then(imgToSend => {
 				msg.channel.send({